Move layout stuff into own folder
